Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,18 +1,8 @@
-// #include <stdio.h>
-
-// char* welcome() {
-//     return "Welcome to Code Radar!";
-// }
-
-// int main() {
-//     printf("%s", welcome());
-//     return 0;
-// }
 #include <stdio.h>
 int main()
 {
-    float a;
-    scanf("%f",&a);
+    double a;
+    scanf("%1f",&a);
     printf("You entered: %1f",a);
     return 0;
 }
